move code aroud now only the server calls set_keyboard
