+++ po/POTFILES.in
@@ -34,11 +34,9 @@ subiquity/common/api/tests/test_server.py
 subiquity/common/errorreport.py
 subiquity/common/filesystem.py
 subiquity/common/__init__.py
-subiquity/common/keyboard.py
 subiquity/common/serialize.py
 subiquity/common/tests/__init__.py
 subiquity/common/tests/test_filesystem.py
-subiquity/common/tests/test_keyboard.py
 subiquity/common/tests/test_serialization.py
 subiquity/common/types.py
 subiquitycore/async_helpers.py
@@ -132,6 +130,7 @@ subiquity/server/controllers/refresh.py
 subiquity/server/controllers/reporting.py
 subiquity/server/controllers/snaplist.py
 subiquity/server/controllers/ssh.py
+subiquity/server/controllers/tests/test_keyboard.py
 subiquity/server/controllers/userdata.py
 subiquity/server/controllers/zdev.py
 subiquity/server/dryrun.py
